Import keyboards from inline module
